+++ src/common/Smi.Common/Helpers/DateTimeProvider.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+
+namespace Smi.Common.Helpers
+{
+    public class DateTimeProvider
+    {
+        public virtual DateTime UtcNow() => DateTime.UtcNow;
+    }
+}