+++ src/common/Smi.Common/Options/GlobalOptions.cs
@@ -7,10 +7,10 @@ using System.Reflection;
 using System.Text;
 using Dicom;
 using FAnsi.Discovery;
+using JetBrains.Annotations;
 using Rdmp.Core.DataLoad.Engine.Checks.Checkers;
 using Rdmp.Core.Repositories;
 using Rdmp.Core.Startup;
-using ReusableLibraryCode.Annotations;
 using Smi.Common.Messages;
 using YamlDotNet.Serialization;
 using DatabaseType = FAnsi.DatabaseType;
@@ -110,7 +110,8 @@ namespace Smi.Common.Options
             return GenerateToString(this);
         }
     }
-
+    
+    [UsedImplicitly]
     public class IsIdentifiableOptions : ConsumerOptions
     {
         /// <summary>
@@ -198,8 +199,9 @@ namespace Smi.Common.Options
         /// Optional, if set then your <see cref="SwapperType"/> will be wrapped and it's answers cached in this Redis database.
         /// The Redis database will always be consulted for a known answer first and <see cref="SwapperType"/> used
         /// as a fallback.
+        /// See https://stackexchange.github.io/StackExchange.Redis/Configuration.html#basic-configuration-strings for the format.
         /// </summary>
-        public string RedisHost { get; set; }
+        public string RedisConnectionString { get; set; }
 
         public override string ToString()
         {
@@ -267,9 +269,8 @@ namespace Smi.Common.Options
             {
                 var opt = (FileReadOption)Enum.Parse(typeof(FileReadOption), FileReadOption);
 
-                //TODO(Ruairidh 2019-08-28) Monitor the status of this
                 if (opt == Dicom.FileReadOption.SkipLargeTags)
-                    throw new ApplicationException("SkipLargeTags option is currently disabled due to issues in fo-dicom. See: https://github.com/fo-dicom/fo-dicom/issues/893");
+                    throw new ApplicationException("SkipLargeTags is disallowed here to ensure data consistency");
 
                 return opt;
             }
@@ -328,9 +329,12 @@ namespace Smi.Common.Options
     public class CohortPackagerOptions
     {
         public ConsumerOptions ExtractRequestInfoOptions { get; set; }
-        public ConsumerOptions ExtractFilesInfoOptions { get; set; }
-        public ConsumerOptions AnonImageStatusOptions { get; set; }
-        public uint JobWatcherTickrate { get; set; }
+        public ConsumerOptions FileCollectionInfoOptions { get; set; }
+        public ConsumerOptions AnonFailedOptions { get; set; }
+        public ConsumerOptions VerificationStatusOptions { get; set; }
+        public uint JobWatcherTimeoutInSeconds { get; set; }
+        public string ReporterType { get; set; }
+        public string NotifierType { get; set; }
 
         public override string ToString()
         {
@@ -556,12 +560,6 @@ namespace Smi.Common.Options
         public string FatalLoggingExchange { get; set; }
         public string RabbitMqControlExchangeName { get; set; }
         public bool ThreadReceivers { get; set; }
-
-        public bool Validate()
-        {
-            return RabbitMqHostPort > 0 &&
-                   !string.IsNullOrWhiteSpace(RabbitMqVirtualHost);
-        }
 
         public override string ToString()
         {