+++ tests/common/Smi.Common.Tests/MicroserviceTester.cs
@@ -39,7 +39,7 @@ namespace Smi.Common.Tests
         {
             CleanUpAfterTest = true;
 
-            _adapter = new RabbitMqAdapter(rabbitOptions, "TestHost");
+            _adapter = new RabbitMqAdapter(rabbitOptions.CreateConnectionFactory(), "TestHost");
 
             _factory = new ConnectionFactory
             {
@@ -174,7 +174,7 @@ namespace Smi.Common.Tests
             foreach (MicroserviceHost host in StopOnDispose)
                 host.Stop("MicroserviceTester Disposed");
 
-            _adapter.Shutdown();
+            _adapter.Shutdown(RabbitMqAdapter.DefaultOperationTimeout);
         }
 
         /// <summary>